requirements.txt
lxml
markdown
python-frontmatter
pyupgrade
